Added Scala example using spark.ml and dataframes
