--- movie-recommendations/movies.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:832b1464a18039e2cb6420ac685b5c026fb717fd9cc94ccc6efedf35fe0ea3cb
-size 25422486
