+++ movie-recommendations/submit-python.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+cd ${DIR}
+
+spark-submit \
+  --master local[*] \
+  --packages org.mongodb.spark:mongo-spark-connector_2.10:0.4 \
+  ${DIR}/src/main/python/movie-recommendations.py
